backend/requirements.txt
asgiref==3.8.1
Django==4.2.13
django-cors-headers==4.4.0
python-decouple==3.8
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
djangorestframework==3.15.2